Add project root flag to Rascal compile command (#1)
This additional projectRoot parameter helps in abbreviating error messages to fit on the screen by removing the common prefix up until the root of each project. It has no other known effects to the best of my knowledge. It is always passed by `mvn compile` to the `RascalCompile` shell command.

The projectRoot is passed to the message printer along with the list of messages. Files that do not fit into the project will be printed with their absolute paths.